email/provider/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.3.0
python-dotenv==1.0.0
pydantic==2.9.2
httpx==0.24.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
msal==1.24.1
requests==2.31.0
python-multipart==0.0.6
email-validator==2.1.0
cryptography==43.0.3
